requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pdf2image==1.16.3
Pillow==10.1.0
opencv-python==4.8.1.78
PyPDF2==3.0.1
openai==1.3.7
requests==2.31.0
python-dotenv==1.0.0
aiofiles==23.2.0
aiohttp==3.9.1
jinja2==3.1.2
pyttsx3==2.90 